requirements.txt
# =============================================================================
# EMAIL DIGEST ASSISTANT - PYTHON DEPENDENCIES
# =============================================================================

# Gmail API Dependencies
# ----------------------
google-auth==2.23.4                 # Core Google authentication
google-auth-oauthlib==1.1.0         # OAuth 2.0 authentication flow
google-auth-httplib2==0.1.1         # HTTP transport for Google APIs
google-api-python-client==2.108.0   # Gmail API client library

# Web Framework 
# flask==2.3.3

# Templating 
# jinja2==3.1.2

# Deployment 
# gunicorn==21.2.0

transformers==4.35.2            # Hugging Face transformers for AI models
torch==2.2.2                    # PyTorch deep learning framework
torchvision==0.17.2            # PyTorch computer vision
torchaudio==2.2.2              # PyTorch audio processing
spacy==3.7.2                   # Industrial NLP library
nltk==3.8.1                    # Natural Language Toolkit
textblob==0.17.1               # Simple text processing
scikit-learn==1.3.2            # Machine learning utilities
sentence-transformers==2.2.2    # Semantic similarity models
numpy==1.26.4                  # Numerical computing
pandas==2.1.4                  # Data manipulation
matplotlib==3.8.0              # Plotting and visualization